feat: Verify the poset is acyclic.py
Patch to catch
```H = nx.DiGraph()
H.add_edges_from([(0, 1), (1, 0)])  # cycle
P = hasse.PoSet(H)```

If we started with PoSet, the construction error was already detected,
therefore I removed half of the added verification. The problem is when
we started with a graph, this error now is detected.

Note that we do not apply transitive reduction during add or
from_chains, instead we reduce when we create a poset. Transitive
reduction already checks for acyclic input, so this solves both
problems.

closes: #1